// HumanResourcesReact/src/components/models/enum.tsx
// src/components/models/enum.ts

export enum SubscriptionPlan {
    MONTLY = "MONTLY",
    ANNUALY = "ANNUALY"
}

export enum EmployeeLimitLevel {
    LEVEL_1 = "EMPLOYEE_LIMIT_LEVEL_1",
    LEVEL_2 = "EMPLOYEE_LIMIT_LEVEL_2",
    LEVEL_3 = "EMPLOYEE_LIMIT_LEVEL_3",
    LEVEL_4 = "EMPLOYEE_LIMIT_LEVEL_4"
}

export enum LeaveType {
    ANNUAL = "ANNUAL",
    SICK = "SICK",
    MATERNITY = "MATERNITY",
    OTHER = "OTHER"
}